requirements.txt
# RESEARCH: intentionally weak pins. Install only in isolated sandboxes.
# SCA: urllib3, requests, older crypto lib; SAST: app code (see README)
# Stack supports current CPython; use requirements-sca-legacy.txt in Docker 3.10 for older SCA.

Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.3
itsdangerous==2.1.2
blinker==1.8.2
click==8.1.7

PyYAML==6.0.1
# Still behind current; not latest (SCA can flag). Pair must match modern CPython.
requests==2.31.0
urllib3==2.0.7
Pillow==10.4.0
ecdsa==0.11
lxml==5.3.0
# 3.5+ for Python 3.12+ importlib; still a dated line for SCA
Markdown==3.6.0
cryptography==42.0.0

# SCA: additional runtime call sites (see /sca/run?k=… and bookstore/sinks/sca_stubs.py)
# httpx (async) — older 0.2x lines have CVEs; 0.27+ stays pip-installable on modern CPython
httpx==0.27.0
paramiko==2.12.0
redis==4.5.4
pycryptodomex==3.9.4
python-jose==3.2.0
protobuf==3.20.3
# ujson: C-backed JSON, older lines common SCA noise (wheels on common platforms)
ujson==5.4.0
MarkupSafe==2.1.3
